WIP #2

+++ includes/op.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   op.h                                               :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jdouniol <jdouniol@student.42.fr>          +#+  +:+       +#+        */
+/*   By: pcarles <pcarles@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2013/10/04 11:33:27 by zaz               #+#    #+#             */
-/*   Updated: 2019/03/13 17:23:41 by jdouniol         ###   ########.fr       */
+/*   Updated: 2019/03/14 16:39:14 by pcarles          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -127,7 +127,7 @@ typedef struct			s_process
 	char				*file_path;
 	char				name[PROG_NAME_LENGTH + 1];
 	char				comment[COMMENT_LENGTH + 1];
-	size_t				program_counter;
+	uint16_t			program_counter;
 	uint8_t				carry;
 	int32_t				registers[REG_NUMBER]; // c est bien ici qu on doit set les valeurs des registres
 	size_t				live_counter;
